Enable automatic migration of /etc contents (overlays)
Create new /etc overlay for each snapshot and mount the stack
matching the current snapshot; this will enable RPM scripts which are
trying to modify existing configuration files to be able to do so now
(up until now they just didn't see the files in the overlay).
Each snapshot will boot the matching overlay file system stack. Files
modified on the running system will also be visible in the new
snapshot (as long as the file wasn't also modified by installation
scripts - in this case the modified file of the newest overlay would
win).
Note that this will approach create a large stack of overlays over
time.