Implement merging changed /etc files
Previously the overlayfs setup automatically made /etc changes done
to the currently running system between creating the snapshot and
rebooting the system visible in the new snapshot.

By removing the /etc overlays this functionality now has to be
implemented manually now. To do so a reference snapshot is created
when the snapshot is taken, which can then be compared against when
booting the new snapshot against the /etc state of the current state
(i.e. these are the files that have been changed during the update)
and the /etc state of the previous snapshot (i.e. these files have
been changed after the snapshot creating and before the reboot).